Redirect some views in read-only mode, hide links to these views.
Read-only mode is a setting an admin can enable to prevent writes to the
database. This can be used when the site is under maintenence or being
upgraded. This commit adds a decorator and class-based view mixin to disable
certain views when read-only mode is enabled..

"New Review Request" and "My Accounts" now redirect to a Read-Only page and
"Register" redirects to "Log-in". Buttons to these pages are removed for users
affected in read-only mode.

Testing Done:
- Check that accessing New Review Request and My Accounts redirects to 503
  page logged in as non-admin in read-only mode.
- Check that accessing Register when logged in or not redirects to the log-in
  page.
- Check that links to New Review/My Accounts/Register are removed for
  non-admins/unauthenticated users.
- Check that submitting Register form when in read-only mode does not do any
  writes.
- Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/8811/