Improve the coloring of primary buttons.
Primary buttons (such as the Log In button) have been a sort of dark
blue/grey in Review Board for a while now. RBCommons replaced this with
a brighter blue button, which feels more fresh and clean, and is a
better match for the rest of our UI.

This change adopts that button color (basically -- slight change in the
gradient) for Review Board.

Testing Done:
Tested the buttons in the UI. We've actually been testing this style for
years on RBCommons.

Made sure the hover effect still worked.

Reviewed at https://reviews.reviewboard.org/r/9255/