Add read-only banner under navbar.
Read-only mode is a setting an admin can enable to prevent writes to the
database. This can be used when the site is under maintenence or being
upgraded. This commit adds a banner under the nav bar when the site is in
read-only mode.

Testing Done:
- Checked that the banner appears in read-only mode.
- Checked that the banner does not appear when not in read-only mode.

Reviewed at https://reviews.reviewboard.org/r/8812/