Fixed the Update -> Add File review request action not responding to clicks.
In the Review Request screen, clicking "Update" -> "Add File" would not do
anything.

This was because the generic menu element's click handler is registered before
the "Add File" element's handler. As a result, the Update element's click
handler is running first and eating up the click event, thus causing the upload
file dialog to never appear.

To fix this, the `onUploadFileClicked()` function was moved from
`ReviewablePageView` to `ReviewRequestEditorView`, and the element's handler
registered in the `setupActions()` function. The `.has-menu` handler was moved
to `ReviewablePageView`, better respecting the ownership structure of the page.
This changes the order in which the click events are fired, allowing the
`onUploadFileClicked()` function to actually fire in response to the "Add File"
element being clicked.

Furthermore, the `has-menu` handler was moved from `ReviewRequestEditorView` to
`ReviewablePageView`.

After the fix, the upload file dialog appears as intended.

Testing Done:
- Ran JSTests.
- Tested manually by clicking Update -> Add File, where the upload file dialog
  appears. This was done on both computer and mobile browsers.

Bugs closed: 4592

Reviewed at https://reviews.reviewboard.org/r/9207/