Update an infobox and datagrid column for issue verification.
This enhances two parts of the UI for the new issue verification
feature: The review request infobox and the issues/Ship It dashboard
column.

The review request infobox's status pane now includes the number of
issues pending verification. This is only shown if there are issues
pending verification.

The issues column in the dashboard was previously updated to include the
count in parenthesis, but it wasn't immediately clear what this count
was for. The column has been updated to show the issue verifications
icon alongside the count when there's at least one issue pending
verification. Under the hood, that code has also been updated to be more
centralized and to more easily support additional counters in the
future.

Testing Done:
Tested the infobox with review requests with and without issues pending
verification. Saw the icon and count when there were such issues.

Tested all states of the dashboard column.

Reviewed at https://reviews.reviewboard.org/r/9254/