Fix the placement of the infobox on the dashboard.
On narrower (but still common) window widths, the review request infobox
was sometimes appearing on the left-hand side, right where the mouse is,
causing problems with mousing around the dashboard. This was due to the
infobox placement priorities being backwards, with bad reasoning for why
they were that way.

The new order is "Ll", prioritizing placement so that the left of the
infobox is relative to the left of the summary (pushed out toward the
right a bit by setting the distances), and falling back to placing the
right-hand side to the left of the summary (when the column is more
toward the right of the dashboard).

Testing Done:
Tested the dashboard with wide widths, narrow widths, and different
locations for the Summary column. It appeared to consistently place the
infobox where I expected it to be.

Reviewed at https://reviews.reviewboard.org/r/9277/