Enable the rbintegrations extension by default.
The rbintegrations extension is a core part of the Review Board 3.0
feature set, and we want to reduce the number of hoops required to start
using integrations. This change marks the extension as being
auto-enabled by default when first registered by the extension manager,
meaning that 3.0 users will automatically have it enabled. Beta users
will have to enable it manually.

Testing Done:
Removed the registration for rbintegrations and re-scanned for extensions.
It found rbintegrations and auto-enabled it.

Reviewed at https://reviews.reviewboard.org/r/9274/