Handle new, changed and deleted symlinks in git.
Mark up file diffs that add, change or remove a symlink so that a label can be
shown in the UI to indicate that the file is a symlink.

Also normalize a git patch containing a symlink to make the file appear as a
regular file, otherwise patch fails to apply the diff, complaining about the
file not being a symlink.

Testing Done:
Unit test.

Tested with a patch adding a symlink and changing another symlink. The files
were properly marked with " (symlink)" in the UI.

Reviewed at https://reviews.reviewboard.org/r/8785/