Modernize and document all extension hooks.
Starting in Djblets 0.10, extension hooks are expected to override the
`initialize` method instead of `__init__`, and no longer need to call
the parent `shutdown`. This was intended both to slim down the code for
hooks and to allow for dynamically disabling/re-enabling hooks and to
optionally start hooks in a disabled state.

This change updates all of our hooks to use the new `initialize` and
`shutdown` behavior, and to make use of base registry hooks where
possible. It also enhances the documentation, providing better guidance
for the kinds of values that each hook expects.

Testing Done:
Unit tests pass.

Tested that my installed extensions didn't outright break.

Reviewed at https://reviews.reviewboard.org/r/9285/