Fix a regression in supporting Power Pack's PDF review.
The cleanup of the JavaScript page models/views moved
`RB.ReviewabePageView.reviewRequest` into the model as an attribute.
This breaks Power Pack, which expects this attribute to exist.

While we will want to update Power Pack to optimistically use the new
attribute, we don't want to break existing installs of Power Pack when
people upgrade. This restores the attribute for compatibility with a
comment explaining the situation.

Testing Done:
Tested that PDF document review worked again.

Reviewed at https://reviews.reviewboard.org/r/9289/