Improve the infobox display when moving the mouse in/out of the box.
Infoboxes are no longer so eager to hide when moving the mouse out and
back onto a row, and are less likely to disappear when trying to mouse
over the infobox. The delays have been increased a bit to keep infoboxes
on-screen longer during mouse movement and to not pop up so quickly when
generally moving the mouse down the datagrid.

There's also a fix for an extra timer that wasn't cleared on mouseleave
events, and one that wasn't cleared when mousing over the target
element.

This does a lot to improve the interaction on the Dashboard.

Testing Done:
Tested moving the mouse back-and-forth between a target row and the
review request infobox, including making a path over neighboring rows.
Verified that the infobox remained on-screen for a longer period of time
and didn't go away and then come back.

Tested moving the mouse from the infobox to a separate target row. The
infobox remained on-screen but then transitioned to the new contents.

Reviewed at https://reviews.reviewboard.org/r/9301/