Improve documentation for mimetypes and API server capabilities.
This modernizes some of the documentation for mimetypes and API server
capabilities, based on the work done on the extension hooks. It adds
documentation for arguments and return types and adds a file-level
docstring for the code reference index.

Testing Done:
Built the docs and looked through them. No build errors.

Reviewed at https://reviews.reviewboard.org/r/9288/