Fix display of floating anchors for the initial status updates.
The floating anchors beside the entries and comments were partially or
entirely off-screen for the initial status updates entry, due to
assumptions that they'd be placed alongside entries with avatars. This
change fixes up those assumptions by setting appropriate offsets based
on whether the entry has an avatar.

While there's only so much space alongside the initial status updates
entry, there's just enough room for the link and a couple of pixels of
space.

There's also an adjustment being applied to floating anchors for
replies to compensate for the border used for highlighting new replies
(which is always present, though transparent if a reply is not new).
This keeps the anchors aligned perfectly.

Testing Done:
Tested the anchor positioning for the initial status updates entry
and on other types of entries, both for top-level comments and replies.
Checked positioning to make sure all anchors on a given entry were
pixel-aligned.

Reviewed at https://reviews.reviewboard.org/r/9307/