Flesh out the documentation on Ship It-related functionality.
Our documentation on marking reviews with Ship It! was a bit stale and
didn't cover some of the modern features of approval. It's been
rewritten to better cover the methods for marking Ship It!, what a
"Fix It, then Ship It!" review means, and how to revoke a Ship It!. This
all comes with screenshots, helping to give people a visual on what this
all looks like.

Testing Done:
Built the docs and read through them. Checked for build failures and
spelling errors.

Made sure the normal and high-res images aligned properly when
transitioning without causing any visual jump.

Reviewed at https://reviews.reviewboard.org/r/9295/