Add docs on changing ownership and touch upon writing custom fields.
This updates the Review Request Fields documentation to add a section on
the Submitter field, talking about its purpose and how in 3.0 the owner
can be reassigned.

It also adds a small section toward the bottom that mentions that the
list of fields can be changed by writing an extension, linking to the
new documentation for this.

Testing Done:
Built the docs and read through them, checking for build errors and
broken links.

Reviewed at https://reviews.reviewboard.org/r/9310/