Make General Comment creation in the UI more clear.
The introduction of General Comments added a button to the Review Dialog
and to the review request's action bar saying "Add Comment." This label
doesn't make clear that it's a General Comment being added, and could
easily be confused as the way one is supposed to do code review
(similarly to how some people have been confused in the past trying to
put their review in the Header field).

This change renames these both to "Add General Comment," helping to
clarify just a bit. It's a bit longer, but there's plenty of room on the
review request's action bar and in the Review Dialog for this, and it
should hold us over until the new review UI is ready.

It also moves the button in the Review Dialog over to the left of the
dialog, separating it from buttons that impact the overall
review/dialog.

Testing Done:
Tested that both of these buttons still work.

Tested the Review Dialog's button placement in desktop and mobile modes.

Reviewed at https://reviews.reviewboard.org/r/9313/