Add CSS rules for segmented buttons.
This extends the button styles to enable support for segmented buttons.
These are elements that look like a single button, but have several
independent components, such as a separate dropdown arrow that behaves
differently from the main button label. Each segment has its own
background, for hover/click effects. They're all still bound to the
visual style of the outer button, though.

This is similar to `SplitButtonView`, and in fact will be used as the
base for that in an upcoming change in order to simplify things and
create a better visual consistency.

Testing Done:
Tested this along with an upcoming change for `SplitButtonView`. The
styles were tested on macOS (Chrome, Firefox, Safari) and on Windows
(IE, Edge, Firefox).

Made sure all existing button styles still looked as they did before.

Reviewed at https://reviews.reviewboard.org/r/9316/