Update the documentation on publishing reviews.
The documentation has been extended to cover how to publish a review
only to the submitter, and how that impacts notifications for the
review. It also adds screenshots showing the Publish Review button with
and without the To Submitter Only option on the Publishing Reviews page,
and shows the screenshot of the banner on the Draft Banner page.

Testing Done:
Built the docs. No build errors or spelling errors noted.

Checked the screenshots to make sure the contents were properly
pixel-aligned for Retina/non-Retina modes.

Reviewed at https://reviews.reviewboard.org/r/9318/