Add docs providing an overview of reviews and the types of comments.
This adds a "Reviews and Comments" page for the "Basic concepts" section
of the User Guide. This page covers what a review is, how it's created
and managed, and the different components of a review. It goes into each
type of comment (adding new documentation on General Comments in the
meantime) and links to different sections to learn more.

Testing Done:
Built the docs. Didn't see any errors.

Checked for spelling errors.

Verified that other docs could now reference the General Comments section.

Reviewed at https://reviews.reviewboard.org/r/9321/