Add a section to the docs about automated code review.
This currently has a quick description of what status updates are
(including a screenshot of some Review Bot output), and links to the
docs for the CircleCI and Travis CI integrations. In the future we'll
also be able to add docs for Review Bot here.

Testing done:
Built HTML and checked the output.

Reviewed at https://reviews.reviewboard.org/r/9328/