Rename the "to_submitter_only" flag to "to_owner_only".
Review Board 3.0 introduces support for only notifying the owner of a
review request via e-mail whne publishing reviews. This was the
"... to Submitter Only" option in the Publish Review button, and it had
a corresponding `to_submitter_only` flag provided in the API, signals,
and other methods.

In an effort to begin standardizing on "owner" as the terminology
instead of "submitter," this change renames all the options and UI for
this feature. This feature wasn't shipped prior to 3.0, so it's safe to
ignore backwards-compatibility here.

It also improves the "to Submitter Only" labels by better describing the
default behavior. The new labels are "... and only e-mail the owner".
The documentation has been updated accordingly, and now mentions that
other extensions/integrations may also decide to react to this option.

Testing Done:
All Python and JavaScript unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/9323/