Add some documentation on issue verification and update screenshots.
This change adds a small bit about issue verification (including advice
on when to use it), and adds in new screenshots for the comment dialog
and the full comment dialog. One small benefit of this is that the
example comment is now grammatically correct.

Testing done:
Built HTML and checked the output.

Reviewed at https://reviews.reviewboard.org/r/9331/