Replace the "Submitter" condition with an "Owner" condition.
In an effort to begin moving to "owner" as the new terminology for a
review request instead of "submitter," this change renames the condition
choice names and values to use "owner" instead. This will break any
conditions on the Submitter field filed during the beta, but it's a
worthwhile breakage in order to avoid maintaining legacy keys going
forward in conditions.

`ReviewRequest` also gained an `owner` property that aliases
`submitter`, which is used in the condition choices and will be used in
future changes. `ReviewRequestDraft` has such an attribute already, so
this provides a good counterpart, allowing callers to access this value
the same way for both objects, much like how `submitter` is set up in
`ReviewRequestDraft`.

Testing Done:
All unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/9333/