Rename the "Submitter" column to "Owner" in the datagrids.
The datagrid columns for the owner of a review request now say "Owner"
in the column header instead of "Submitter." The actual ID of the column
is the same, so that we don't break any existing users, but they'll see
the new label in the column heading and Edit Columns menu.

Documentation has been updated. Screenshots still need to be updated, but
this will be done separately.

Testing Done:
Verified that the column changed to "Owner," but that nothing else broke.

Built the docs and saw the new text.

Reviewed at https://reviews.reviewboard.org/r/9335/