Replace the "Submitter" field with an "Owner" field.
This replaces the "Submitter" field with an equivalent "Owner" field. It
keeps the same field ID for compatibility, but the class names and
labels are updated.

From a user standpoint, this is purely a presentational change. From a
code standpoint, it's primarily just a search-and-replace of relevant
class names.

Documentation has been updated to reference this field.

Testing Done:
Unit tests passed.

Changed the owner field without any problems. Saw that the history was
recorded.

Built the docs and verified that the field was properly documented.

Reviewed at https://reviews.reviewboard.org/r/9334/