Fix a crash and layout problem when editing the Testing Done field.
The recent work to port `$.fn.inlineEditor` to Backbone missed a
variable update (`_$buttons` to `$buttons`) in the review request layout
code, which could result in a crash and leave the layout in a bad state.
This would manifest as a short text field when starting to edit, and
file attachments overlaying text field when completing the editing.

To reproduce this, the last text field needed to have its height
expanded to the height of the fields on the side when not in edit mode.
This was easily the case if the Description field was empty or only
contained a line or two, and the Testing Done field was being filled
with many lines of content.

The variable fix avoids the crash, allowing the layout code to complete
successfully.

Testing Done:
Reproduced both the short textfield bug and the overlapping file
attachments bug, and saw the crash in both cases. Verified that this
fixed both.

Bugs closed: 4601

Reviewed at https://reviews.reviewboard.org/r/9340/