Update the image review docs to discuss zoom levels.
The image review docs now have a section on changing the zoom level of
the image. These discuss the various zoom levels that can be chosen,
along with the defaults used based on the filename and the image size.

The screenshot of the image review UI has also been updated to include
the new UI with the zoom level.

Testing Done:
Built and viewed the docs. Checked for build and spelling errors.

Checked the new image to ensure pixel alignment.

Reviewed at https://reviews.reviewboard.org/r/9355/