Add docs on file attachment navigation and update screenshots.
This adds a section to the file attachment overview docs covering the
ability to switch file attachments on 3.0, showing a screenshot of a PDF
review UI with the navigation buttons present, one expanded to show the
thumbnail.

It also updates the screenshots on the page for the revision selector.
The new screenshot reflects how the UI looks now (instead of the old one
used when the screenshot was first taken), and now only shows one view
of the slider and text instead of several (helping to simplify it so it
doesn't look like there's multiple sliders and making it easier to keep
it up-to-date). This comes with a new Retina version, which the old one
did not have.

Testing Done:
Built the docs. Checked for build errors, spelling errors, and
visual errors with the screenshots.

Reviewed at https://reviews.reviewboard.org/r/9360/