Add Markdown docs on uploading images, and add new links to sub-sections.
The Markdown documentation now includes a section on uploading images to
fields. This is pretty basic, but touches upon the feature. It's linked
to in the Images section for the Markdown reference.

The main page for the User Manual also improves the Markdown section,
linking to this new sub-section as well as the Basic Syntax, Code
Samples, and Emoji sub-sections.

Testing Done:
Built the docs and checked the links.

Checked spelling.

Reviewed at https://reviews.reviewboard.org/r/9354/