Respond to events on file attachments/screenshots from page load.
Changing the caption on a file attachment/screenshot that was present on
page load would not result in a draft banner appearing until the page
was next reloaded. This happened because we didn't register any of the
file attachments or screenshots that were provided as part of the page
load, preventing their events from being connected with the
`ReviewRequestEditor`'s events.

This change registers existing attachments/screenshots and adds unit
tests to ensure this works and does not regress.

Testing Done:
Edited the caption on a file attachment on a review request without a
draft. Saw that the draft banner appeared, whereas prior to this it
would not. Repeated this test for screenshots.

Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/9366/