Fix text wrapping for replies in status update reviews.
Replies to comments on status updates used standard "pre" wrapping, only
wrapping to the next line if encountering newlines. That's because the
rules governing the wrapping of comment text only applied to comments in
top-level review entries and not to reviews in status updates.

There's now a new rule for the comment text that adds "pre" wrapping.
This will work in any kind of entry that has a review thread and uses
the standard classes.

Testing Done:
Created some long comments on replies to reviews and status update reviews.
Saw that the text was wrapped at the element boundaries.

Reviewed at https://reviews.reviewboard.org/r/9362/