Fix the screenshots for integrations list and conditions.
The Retina screenshots showing the integrations list and configuring
conditions was off by 1 pixel, causing a jump when transitioning from a
low-DPI to a high-DPI screenshot. This change crops the integrations
list to keep things consistent, and shifts the contents of the
conditions screenshot right 1 pixel to fix that one up.

The images have also been optimized, reducing file size by 40%.