Improve the display of repository types in the admin manual.
The index of the administration manual has a list of repository types
and hosting services, which are displayed as a single line with a "|"
separating each entry. This wasn't the best layout for this, and a "|"
was even missing.

This change switches these entries to use Sphinx's `hlist::` directive,
which creates a nice multi-column layout (2 columns for HTML/pickle),
creating a much nicer view for these lists.

Testing Done:
Built the docs. Checked the resulting lists.

Reviewed at https://reviews.reviewboard.org/r/9376/