List available integrations on the admin docs index.
The Integrations page lists all the integrations that can be configured,
acting as the parent for those pages. This change adds some convenient
references to the index of the administration guide that link to the
specific integrations. This helps both to advertise the integrations
available when making a cursory glance through the administration page
and to make it a bit easier to locate the docs for those already
familiar with integrations. These are shown as nested under the link to
the Integrations page.

Testing Done:
Built the docs and tested the links.

Reviewed at https://reviews.reviewboard.org/r/9375/