Fix updating the diff file index when changing pagination.
The diff file index wasn't properly updating when switching pages in a
diff. This was due to some of the work done to move diff loading into
the models. An array was no longer being cleared and the way the file
list was being repopulated resulted in a different event name being
used. This fixes both of those, allowing the file index to update
whenever the list of files changes.

Testing Done:
Changed pages in a diff with lots of files. Saw that the file list
fully repopulated each time, updating the filenames, anchors, and
icons.

Bugs closed: 4606

Reviewed at https://reviews.reviewboard.org/r/9373/