Add documentation on OAuth2
This change adds a new page to the docs for how to use OAuth with the
Review Board API. This includes the specifics of using the UI to
register a new application, as well as the details of the authorization
handshake.

Testing done:
Built HTML, checked output.

Reviewed at https://reviews.reviewboard.org/r/9378/